backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
pgvector>=0.2.5
langchain>=0.1.13
langchain-text-splitters>=0.0.1
sentence-transformers>=2.6.1
google-genai>=0.2.2
asyncpg>=0.29.0
loguru>=0.7.2
langsmith>=0.1.0
PyMuPDF>=1.24.1
pydantic>=2.6.4
python-dotenv>=1.0.0
